Remove the `FastCompString` class and use `std::string` instead + move the `==` operator overload of the removed class to the new function `stringsEqual()`

+++ src/linked_lists.hpp
@@ -21,7 +21,7 @@ public:
     Link(){};
     ~Link()
     {
-        for (int i = 0; i < BASE_ARRAY_LEN; ++i)
+        for (short i = 0; i < BASE_ARRAY_LEN; ++i)
         {
             delete elts[i];
         }
